requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
moviepy>=1.0.3
pydub>=0.25.1
srt>=3.5.3
ffmpeg-python>=0.2.0